+++ src/lib/public_api.ts
@@ -41,7 +41,7 @@ export { PipesModule } from './pipes/pipes.module';
 export { StFilterList } from './pipes/search-filter.pipe';
 
 /**************************** DIRECTIVES ***************************************************/
-export { StInputAdjustable } from './directives/st-input-adjustable';
+export { StInputAdjustable } from './directives/st-input-adjustable/st-input-adjustable';
 export { StMinValidator } from './directives/form/st-min-validator/st-min-validator';
 export { StMaxValidator } from './directives/form/st-max-validator/st-max-validator';